requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
django-jazzmin==3.0.1
django-modeltranslation==0.19.10
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
